extract the tex2svg code into its own class
