+++ tests/_data/mathjax.html
@@ -1,3 +1,4 @@
+<!DOCTYPE html>
 <html>
   <head>
     <title>MathJax node rendering test</title>
